#57 - separate out tradescreen, set up styles folder for components
